VideoShaders: rename VideoFilterShader to VideoFilterShaderGL
