+++ xbmc/cores/VideoPlayer/VideoRenderers/LinuxRendererGL.cpp
@@ -33,7 +33,7 @@
 #include "settings/MediaSettings.h"
 #include "settings/Settings.h"
 #include "VideoShaders/YUV2RGBShader.h"
-#include "VideoShaders/VideoFilterShader.h"
+#include "VideoShaders/VideoFilterShaderGL.h"
 #include "windowing/WindowingFactory.h"
 #include "guilib/Texture.h"
 #include "guilib/LocalizeStrings.h"
